Add page for manifesto (from previous project) to portfolio section.
Add links to main page so that both the image and title of the manifesto
project link to the manifesto page. Add return link to manifesto page.

Check HTML and CSS for new page using WC3 Validators.
Test all links in Chrome, Firefox, Internet Explorer, and Edge to verify
that they work properly. Check views in all 4 browsers to verify that it
renders decently in each of them.